Early main07-1.rs; fix V f64 default change

+++ rtiow/V/A09-2.v
@@ -55,7 +55,7 @@ struct MLambertian {
 struct MMetal {
 	scattercb ScatterCallback = ScatterCallback(cb_scatter_metal)
 	albedo vec.Vec3
-	fuzz f32 = 0.
+	fuzz f32 = f32(0.)
 }
 
 union Material {
